requirements.txt
fastapi[standard]
pydantic-settings
python-jose
python-dotenv
sqlalchemy[asyncio]
websockets
asyncpg
alembic